luci-app-ddns: make param opt, optional, add myself as maintainer
Currently these params should be optional but in the current script then are required for any configuration to save. We add an "optional" boolean value to the validate function so that we can make this pram optional as it should be. Also adds me as maintainer

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>
(cherry picked from commit 09c3a8385e2176fc8a4f934e592f4089b662a1c1)